splitting validation by blacklist and log email to separate services
