+++ Mail/Transport.php
@@ -9,33 +9,26 @@
 namespace Mygento\Smtp\Mail;
 
 use Closure;
-use Magento\Framework\Exception\LocalizedException;
 use Magento\Framework\Exception\MailException;
-use Magento\Framework\Mail\Address;
 use Magento\Framework\Mail\EmailMessageInterface;
 use Magento\Framework\Mail\TransportInterface;
-use Mygento\Smtp\Api\Data;
-use Mygento\Smtp\Api\LogRepositoryInterface;
 use Mygento\Smtp\Model\Config;
-use Mygento\Smtp\Model\Source\Status;
-use Psr\Log\LoggerInterface;
+use Mygento\Smtp\Model\Mail\Logger;
+use Mygento\Smtp\Model\Mail\Validator;
 
-/**
- * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
- */
 class Transport
 {
     public function __construct(
-        private LogRepositoryInterface $repo,
-        private Data\LogInterfaceFactory $factory,
+        private Logger $mailLogger,
+        private Validator $blackListValidator,
         private Config $config,
-        private LoggerInterface $logger
     ) {
     }
 
     /**
      * @param TransportInterface $subject
      * @param Closure $proceed
+     * @throws MailException
      * @return void
      * @SuppressWarnings(PHPMD.UnusedFormalParameter)
      */
@@ -46,107 +39,14 @@ class Transport
 
             return;
         }
+        /** @var EmailMessageInterface $message */
         $message = $subject->getMessage();
-        if ($this->validateBlacklist($message)) {
+
+        if (!$this->blackListValidator->isValid($message)) {
             return;
         }
 
-        /** @var Data\LogInterface $entity */
-        $entity = $this->factory->create();
-
-        try {
-            $this->fillEntity($entity, $message);
-            $proceed();
-
-            $entity->setStatus(Status::STATUS_SUCCESS);
-            $this->repo->save($entity);
-        } catch (MailException $e) {
-            $entity->setError($e->getPrevious()->getMessage());
-            $entity->setStatus(Status::STATUS_ERROR);
-
-            throw $e;
-        } catch (LocalizedException $e) {
-            $entity->setStatus(Status::STATUS_ERROR);
-            $entity->setError($e->getMessage());
-            $this->logger->error($e->getMessage(), ['exception' => $e]);
-        } finally {
-            $this->repo->save($entity);
-        }
-    }
-
-    private function validateBlacklist(EmailMessageInterface $message): bool
-    {
-        $result = false;
-        $blacklist = $this->config->getBlacklist();
-
-        if (!$blacklist) {
-            return false;
-        }
-
-        $recipient = $this->getRecipient($message);
-        $patterns = array_unique(explode(PHP_EOL, $blacklist));
-        foreach ($patterns as $pattern) {
-            try {
-                if (preg_match($pattern, $recipient)) {
-                    $result = true;
-                    break;
-                }
-            } catch (\Exception $e) {
-                // Ignore validate if the pattern is error
-                continue;
-            }
-        }
-
-        return $result;
-    }
-
-    private function getRecipient(EmailMessageInterface $message): string
-    {
-        $emails = [];
-        if ($message->getTo()) {
-            foreach ($message->getTo() as $address) {
-                $emails[] = $address->getEmail();
-            }
-        }
-
-        return implode(',', $emails);
-    }
-
-    private function formatList(iterable $list): string
-    {
-        return implode(
-            ',',
-            array_map(fn (Address $address) => $address->getEmail(), $list)
-        );
-    }
-
-    private function fillEntity(Data\LogInterface $entity, EmailMessageInterface $message): Data\LogInterface
-    {
-        if ($message->getSender()) {
-            $entity->setSender(
-                $message->getSender()->getName() . ' <' . $message->getSender()->getEmail() . '>'
-            );
-        } elseif (is_iterable($message->getFrom()) && count($message->getFrom()) > 0) {
-            $f = $message->getFrom();
-            $from = reset($f);
-            $entity->setSender(
-                $from->getName() . ' <' . $from->getEmail() . '>'
-            );
-        }
-
-        $entity->setRecipient($this->formatList($message->getTo()));
-        $entity->setSubject($message->getSubject());
-        $messageBody = quoted_printable_decode($message->getBodyText());
-        $content = htmlspecialchars($messageBody);
-        $entity->setContent($content);
-
-        if (is_iterable($message->getCc()) && count($message->getCc()) > 0) {
-            $entity->setCc($this->formatList($message->getCc()));
-        }
-        if (is_iterable($message->getBcc()) && count($message->getBcc()) > 0) {
-            $entity->setBcc($this->formatList($message->getBcc()));
-        }
-
-        return $entity;
+        $this->mailLogger->logMessage($message);
+        $proceed();
     }
 }
